requirements.txt
streamlit==1.38.0
google-generativeai==0.5.0
gTTS==2.5.3
pandas==2.2.2
PyPDF2==3.0.1
openpyxl==3.1.5
